Chore MVVM Architecture
