+++ src/screens/Internal/styles.ts
@@ -2,7 +2,6 @@ import styled from 'styled-components/native';
 
 export const Container = styled.View`
   flex: 1;
-  justify-content: center;
-  align-items: center;
   background-color: #f1f1f1;
+  justify-content: center;
 `;
